cities fragment updated

+++ data/src/main/java/ru/voodoo420/data/models/Coord.kt
@@ -1,6 +1,6 @@
 package ru.voodoo420.data.models
 
 data class Coord(
-    val lat: Double,
-    val lon: Double
+    val lat: Float,
+    val lon: Float
 )
