+++ data/src/main/java/ru/voodoo420/data/models/Rain.kt
@@ -1,5 +1,5 @@
 package ru.voodoo420.data.models
 
 data class Rain(
-    val `3h`: Double
+    val `3h`: Float
 )
