+++ data/src/main/java/ru/voodoo420/data/models/Wind.kt
@@ -2,5 +2,5 @@ package ru.voodoo420.data.models
 
 data class Wind(
     val deg: Int,
-    val speed: Double
+    val speed: Float
 )
