+++ domain/src/main/java/ru/voodoo420/domain/entities/CurrentWeather.kt
@@ -1,4 +1,4 @@
 package ru.voodoo420.domain.entities
 
-data class CurrentWeather(val city: String, val temperature: Double, val icon: String,
-                          val humidity: Int, val wind: Double, val minTemp: Double, val maxTemp: Double)
+data class CurrentWeather(val city: String, val temperature: Float, val icon: String,
+                          val humidity: Int, val wind: Float, val minTemp: Float, val maxTemp: Float)
