+++ domain/src/main/java/ru/voodoo420/domain/entities/ForecastUnit.kt
@@ -1,3 +1,3 @@
 package ru.voodoo420.domain.entities
 
-data class ForecastUnit(val date: String, val temperature: Double, val icon: String)
+data class ForecastUnit(val date: String, val temperature: Float, val icon: String)
